Add a new TestRunMonitorWorker
This renames the MonitorTestRunWorker, it monitors the running job via
the github api, it then sends a message to slack with the result
